+++ EnumDef.h
@@ -1,0 +1,6 @@
+enum Mode : int
+{
+    CLOSEDLOOP = 0,
+    OPENLOOP = 1,
+    SLOWCONTROL = 2
+};
