+++ Interrupts.h
@@ -1,0 +1,7 @@
+void enable_Switch();
+
+void BUTTONINTERRUPT();
+
+void enable_Switch();
+
+int getMode();
